+++ test/sampler_test.py
@@ -133,8 +133,8 @@ class TestDynesty(unittest.TestCase):
     def setUp(self):
         self.likelihood = MagicMock()
         self.priors = bilby.core.prior.PriorDict()
-        self.priors['a'] = bilby.core.prior.Prior(periodic_boundary=True)
-        self.priors['b'] = bilby.core.prior.Prior(periodic_boundary=False)
+        self.priors['a'] = bilby.core.prior.Prior(boundary='periodic')
+        self.priors['b'] = bilby.core.prior.Prior(boundary='reflective')
         self.sampler = bilby.core.sampler.Dynesty(self.likelihood, self.priors,
                                                   outdir='outdir', label='label',
                                                   use_ratio=False, plot=False,
@@ -156,7 +156,7 @@ class TestDynesty(unittest.TestCase):
                         logl_max=np.inf, add_live=True, print_progress=True, save_bounds=False,
                         walks=20, update_interval=600, print_func='func')
         self.sampler.kwargs['print_func'] = 'func'  # set this manually as this is not testable otherwise
-        self.assertListEqual([0], self.sampler.kwargs['periodic'])  # Check this separately
+        self.assertListEqual([0, 1], self.sampler.kwargs['periodic'])  # Check this separately
         self.sampler.kwargs['periodic'] = None  # The dict comparison can't handle lists
         for key in self.sampler.kwargs.keys():
             print(key)
@@ -165,7 +165,7 @@ class TestDynesty(unittest.TestCase):
         self.assertDictEqual(expected, self.sampler.kwargs)
 
     def test_translate_kwargs(self):
-        expected = dict(bound='multi', sample='rwalk', periodic=[0], verbose=True,
+        expected = dict(bound='multi', sample='rwalk', periodic=[0, 1], verbose=True,
                         check_point_delta_t=600, nlive=1000, first_update=None,
                         npdim=None, rstate=None, queue_size=None, pool=None,
                         use_pool=None, live_points=None, logl_args=None, logl_kwargs=None,
@@ -394,8 +394,8 @@ class TestPymultinest(unittest.TestCase):
     def setUp(self):
         self.likelihood = MagicMock()
         self.priors = bilby.core.prior.PriorDict()
-        self.priors['a'] = bilby.core.prior.Prior(periodic_boundary=True)
-        self.priors['b'] = bilby.core.prior.Prior(periodic_boundary=False)
+        self.priors['a'] = bilby.core.prior.Prior(boundary='periodic')
+        self.priors['b'] = bilby.core.prior.Prior(boundary='reflective')
         self.sampler = bilby.core.sampler.Pymultinest(self.likelihood, self.priors,
                                                       outdir='outdir', label='label',
                                                       use_ratio=False, plot=False,
@@ -458,8 +458,8 @@ class TestRunningSamplers(unittest.TestCase):
             self.x, self.y, self.model, self.sigma)
 
         self.priors = bilby.core.prior.PriorDict()
-        self.priors['m'] = bilby.core.prior.Uniform(0, 5, periodic_boundary=False)
-        self.priors['c'] = bilby.core.prior.Uniform(-2, 2, periodic_boundary=False)
+        self.priors['m'] = bilby.core.prior.Uniform(0, 5, boundary='reflective')
+        self.priors['c'] = bilby.core.prior.Uniform(-2, 2, boundary='reflective')
         bilby.core.utils.check_directory_exists_and_if_not_mkdir('outdir')
 
     def tearDown(self):
@@ -511,8 +511,8 @@ class TestRunningSamplers(unittest.TestCase):
     def test_run_PTMCMCSampler(self):
         _ = bilby.run_sampler(
             likelihood=self.likelihood, priors=self.priors,
-            sampler= 'PTMCMCsampler', Niter=101, burn =2,
-            isave = 100 ,save=False)
+            sampler='PTMCMCsampler', Niter=101, burn=2,
+            isave=100, save=False)
 
 
 if __name__ == '__main__':